Refactor import statements
