+++ src/adam/parametric/model/parametric_factories/parametric_model.py
@@ -3,8 +3,8 @@ import urdf_parser_py.urdf
 from adam.core.spatial_math import SpatialMath
 from adam.model import Joint, Link, ModelFactory, StdJoint, StdLink
 from adam.model.std_factories.std_model import get_xml_string, urdf_remove_sensors_tags
-from adam.model.visuals import normalize_urdf_visual
 from adam.model.std_factories.std_model import URDFModelFactory
+from adam.model.std_factories.urdf_visual import normalize_urdf_visual
 from adam.parametric.model import ParametricJoint, ParametricLink
